requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
